requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tensorboard
torch>=1.7.1
torchvision
tqdm
yapf
lpips
pyiqa # PyTorch Toolbox for Image Quality Assessment
# gdown # supports downloading the large file from Google Drive